requirements.txt
pdfminer.six
beautifulsoup4
